requirements/base.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
bcrypt==4.1.3



chromadb==0.5.20
fastembed==0.3.6
pypdf==5.0.1
langgraph
